requirements.txt
Flask==3.1.1
flask-cors==6.0.1
PyMuPDF==1.26.1
python-docx==1.1.0

# AI/ML packages (optional - for advanced features)
# langchain
# langchain-community
# langchain-text-splitters
# sentence-transformers
# chromadb
# ctransformers
# llama-cpp-python
